Add cron running and start script
This commit adds a new property to the instance config which allows a
cron schedule to be specified for the script.

It also moves the encryption key under which user emails are stored into
something that is pasted in when the app starts.

- Add cron_restart property to ecosystem.config.js.template
- Remove KEY property from app environment in
ecosystem.config.js.template
- Create start.sh, a script that accepts arguments containing the
instance name and environment, then prompts the user to paste the
encryption key, then starts the app, passing the key into the app's
environment.